Update compare options to work with render script (double -> float)

+++ rembrandtAndroid/src/main/java/ly/img/android/rembrandt/RembrandtCompareOptions.java
@@ -8,17 +8,17 @@ import android.graphics.Color;
 
 public class RembrandtCompareOptions {
 
-    public static final double DEFAULT_MAXIMUM_COLOR_DISTANCE = 25.0f;
-    public static final double DEFAULT_MAXIMUM_PERCENTAGE_OF_DIFFERENT_PIXELS = 1.0f;
+    public static final float DEFAULT_MAXIMUM_COLOR_DISTANCE = 25.0f;
+    public static final float DEFAULT_MAXIMUM_PERCENTAGE_OF_DIFFERENT_PIXELS = 1.0f;
     public static final int DEFAULT_COLOR_BITMAP_PIXEL_EQUAL = Color.WHITE;
     public static final int DEFAULT_COLOR_BITMAP_PIXEL_DIFFERENT = Color.BLACK;
 
-    private final double maximumColorDistance;
-    private final double maximumPercentageOfDifferentPixels;
+    private final float maximumColorDistance;
+    private final float maximumPercentageOfDifferentPixels;
     private final int colorForEquality;
     private final int colorForDiversity;
 
-    public RembrandtCompareOptions(final double maximumColorDistance, final double maximumPercentageDifference, final int colorForEquality, final int colorForDiversity) {
+    public RembrandtCompareOptions(final float maximumColorDistance, final float maximumPercentageDifference, final int colorForEquality, final int colorForDiversity) {
         this.maximumColorDistance = maximumColorDistance;
         this.maximumPercentageOfDifferentPixels = maximumPercentageDifference;
         this.colorForEquality = colorForEquality;
@@ -32,11 +32,11 @@ public class RembrandtCompareOptions {
                 DEFAULT_COLOR_BITMAP_PIXEL_DIFFERENT);
     }
 
-    public double getMaximumColorDistance() {
+    public float getMaximumColorDistance() {
         return maximumColorDistance;
     }
 
-    public double getMaximumPercentageOfDifferentPixels() {
+    public float getMaximumPercentageOfDifferentPixels() {
         return maximumPercentageOfDifferentPixels;
     }
